+++ packages/cli/src/devServer.mjs
@@ -169,7 +169,8 @@ export async function runDevServer({ context, merkurConfig, cliConfig }) {
         }
 
         logger.error(error);
-        res.status(500).json({
+        // TODO send html page with preview of error for better DX
+        res.status(error?.status ?? 500).json({
           error: {
             message: `Something is wrong with the @merkur/cli/devServer: ${error.message}`,
             stack: error.stack,